Change telegram_id to BigInteger

+++ app/db/models.py
@@ -26,7 +26,7 @@ class Admin(Base):
     created_at = Column(DateTime, default=datetime.utcnow)
     is_sudo = Column(Boolean, default=False)
     password_reset_at = Column(DateTime, nullable=True)
-    telegram_id = Column(Integer, nullable=True ,default=None)
+    telegram_id = Column(BigInteger, nullable=True ,default=None)
     discord_webhook = Column(String(1024), nullable=True, default=None)
